+++ test/src/Other/IpEndpointTest.cc
@@ -182,7 +182,7 @@ TEST_F(IpEndpointTest, ListBucketspagingTest)
     //list by step
     request.setMaxKeys(2);
     bool IsTruncated = false;
-    int total = 0;
+    size_t total = 0;
     do {
         outcome = Client->ListBuckets(request);
         EXPECT_EQ(outcome.isSuccess(), true);
@@ -191,7 +191,7 @@ TEST_F(IpEndpointTest, ListBucketspagingTest)
         request.setMarker(outcome.result().NextMarker());
         IsTruncated = outcome.result().IsTruncated();
     } while (IsTruncated);
-    EXPECT_EQ(total, 5);
+    EXPECT_EQ(total, 5UL);
 
     //delete all
     OssClient client(Config::Endpoint, Config::AccessKeyId, Config::AccessKeySecret, ClientConfiguration());
